+++ inst/apps/138-icon-fontawesome/app.R
@@ -1,11 +1,20 @@
 library(shiny)
+library(fontawesome)
 
 fa4icons <- c(
-  "scissors",
-  "slack",
-  "arrow-left",
   "eur",
-  "barcode"
+  "mouse-pointer",
+  "rupee",
+  "cogs",
+  "inr"
+)
+
+fa5icons <- c(
+  "innosoft",
+  "apple-alt",
+  "sync",
+  "sort-alpha-up",
+  "redo"
 )
 
 fa6icons <- c(
@@ -13,7 +22,7 @@ fa6icons <- c(
   "socks",
   "bullhorn",
   "lightbulb",
-  "wifi"
+  "worm"
 )
 
 fabicons <- c(
@@ -37,13 +46,16 @@ showIcons <- function(icons) {
 }
 
 ui <- fluidPage(
-  tags$h2("FontAwesome 4 Icons"),
+  tags$h2("Font Awesome 4 Icons"),
   p("The following icons are from FontAwesome 4. They should display properly below."),
   showIcons(fa4icons),
-  tags$h2("FontAwesome 6 Icons"),
+  tags$h2("Font Awesome 5 Icons"),
+  p("The following icons are from FontAwesome 5. They should display properly below."),
+  showIcons(fa5icons),
+  tags$h2("Font Awesome 6 Icons"),
   p("The following icons are from FontAwesome 6. They should display properly below."),
   showIcons(fa6icons),
-  tags$h2("FontAwesome 6 Brand icons"),
+  tags$h2("Font Awesome 6 Brand icons"),
   p("The following icons are from the FontAwesome 6 Brand set. They should display properly below."),
   showIcons(fabicons)
 )